Modulated RenderManager header, updated update func, etc..
IMPORTANT!!! I left SDLRenderObject and RenderResource objects able to
be initialized from the test file for debugging purposes. These should
most likely be turned off once the SceneManager is properly implemented.
