+++ SquadIO/RenderResource.cpp
@@ -1,7 +1,10 @@
 #pragma once
 #include "RenderManager.h"
 
-using namespace std;
+
+RenderResource::RenderResource(){}
+
+RenderResource::~RenderResource(){}
 
 void RenderResource::unload(){
 	if (mSurface){
@@ -15,8 +18,7 @@ void RenderResource::load(){
 	//Can load BMP, GIF, JPEG, LBM, PCX, PNG, PNM, TGA, TIFF, XCF, XPM, XV file formats
 
 	//TODO: variable 'filename' is inherited from the cResource class yet to be implemented
-	//SDL_Surface *tempSurface = IMG_Load(m_Filename.c_str());
-	SDL_Surface *tempSurface = IMG_Load("Duck.png");
+	SDL_Surface *tempSurface = IMG_Load(m_Filename.c_str());
 
 	if (tempSurface){
 		mSurface = tempSurface;
